requirements.txt
attrdict
gensim
matplotlib
nltk
numpy
pandas
regex
scipy
sympy==1.9
tensorboard
tensorboardX
tokenizers
tqdm
transformers
pandas
sentencepiece
ruamel.yaml
word2number
scikit-learn
torch 
torchvision 
torchaudio
tensorflow
tensorflow-datasets
